Implement missing low level features - add Color struct to the engine and refactor it in Lua
