+++ engine/lib/phx/script/libphx.lua
@@ -325,6 +325,13 @@ do -- Transparent Structs
       float uvx;
       float uvy;
     } Vertex;
+
+    typedef struct Color {
+      float r;
+      float g;
+      float b;
+      float a;
+    } Color;
   ]]
 
   libphx.Structs = {
@@ -355,6 +362,7 @@ do -- Transparent Structs
     'Vec4f',
     'Vec4i',
     'Vertex',
+    'Color',
   }
 end
