+++ engine/lib/phx/src/render/color.rs
@@ -1,0 +1,32 @@
+use glam::Vec4;
+
+#[derive(Debug, Clone, PartialEq)]
+#[repr(C)]
+pub struct Color {
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
+    pub a: f32,
+}
+
+impl From<Vec4> for Color {
+    fn from(value: Vec4) -> Self {
+        Self {
+            r: value.x,
+            g: value.y,
+            b: value.z,
+            a: value.w,
+        }
+    }
+}
+
+impl From<Color> for Vec4 {
+    fn from(value: Color) -> Self {
+        Self {
+            x: value.r,
+            y: value.g,
+            z: value.b,
+            w: value.a,
+        }
+    }
+}
