+++ engine/lib/phx/src/render/mod.rs
@@ -3,6 +3,7 @@ mod box_mesh;
 mod box_tree;
 mod bsp;
 mod collision_shape;
+mod color;
 mod cube_face;
 mod cull_face;
 mod data_format;
@@ -43,6 +44,7 @@ pub use box_mesh::*;
 pub use box_tree::*;
 pub use bsp::*;
 pub use collision_shape::*;
+pub use color::*;
 pub use cube_face::*;
 pub use cull_face::*;
 pub use data_format::*;
